+++ backend/src/main/java/application/model/tasks/Task.java
@@ -12,7 +12,7 @@ public class Task implements Completable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String name;
     private String description;
@@ -34,7 +34,7 @@ public class Task implements Completable {
     public Task() {
     }
 
-    public Task(Integer id, String name, String description, String date) {
+    public Task(Long id, String name, String description, String date) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -42,11 +42,11 @@ public class Task implements Completable {
         this.isFinished = false;
     }
 
-    public Task(Integer id, String name, String date) {
+    public Task(Long id, String name, String date) {
         this(id, name, "no description", date);
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
